refactor: move nested classes and enums to their own files
The CiscoRoomOsCodec class is already ~7000 lines by itself, which is incredibly unwieldy. It doesn't need other classes and enums adding lines to it.
