+++ src/CiscoRoomOsCodec.cs
@@ -23,6 +23,7 @@ using PepperDash.Essentials.Devices.Common.Codec;
 using PepperDash.Essentials.Devices.Common.Codec.Cisco;
 using PepperDash.Essentials.Devices.Common.VideoCodec;
 using PepperDash.Essentials.Plugin.CiscoRoomOsCodec.Cameras;
+using PepperDash.Essentials.Plugin.CiscoRoomOsCodec.Enums;
 using PepperDash.Essentials.Plugin.CiscoRoomOsCodec.Interfaces;
 using PepperDash.Essentials.Plugin.CiscoRoomOsCodec.UserInterface.UserInterfaceExtensions;
 using PepperDash.Essentials.Plugin.CiscoRoomOsCodec.UserInterface.WebView;
@@ -32,113 +33,6 @@ using Feedback = PepperDash.Essentials.Core.Feedback;
 
 namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 {
-	#region enums
-	internal enum eCommandType
-	{
-		SessionStart,
-		SessionEnd,
-		Command,
-		GetStatus,
-		GetConfiguration
-	};
-
-	/// <summary>
-	/// Defines the types of external sources that can be connected to a Cisco codec.
-	/// </summary>
-	public enum eExternalSourceType
-	{
-		/// <summary>
-		/// External camera source
-		/// </summary>
-		camera,
-		/// <summary>
-		/// Desktop computer source
-		/// </summary>
-		desktop,
-		/// <summary>
-		/// Document camera source
-		/// </summary>
-		document_camera,
-		/// <summary>
-		/// Media player source
-		/// </summary>
-		mediaplayer,
-		/// <summary>
-		/// Personal computer source
-		/// </summary>
-		PC,
-		/// <summary>
-		/// Whiteboard source
-		/// </summary>
-		whiteboard,
-		/// <summary>
-		/// Other unspecified source type
-		/// </summary>
-		other
-	}
-
-	/// <summary>
-	/// Defines the operational modes of external sources connected to a Cisco codec.
-	/// </summary>
-	public enum eExternalSourceMode
-	{
-		/// <summary>
-		/// External source is ready and available for use
-		/// </summary>
-		Ready,
-		/// <summary>
-		/// External source is not ready for use
-		/// </summary>
-		NotReady,
-		/// <summary>
-		/// External source is hidden from the user interface
-		/// </summary>
-		Hidden,
-		/// <summary>
-		/// External source is in an error state
-		/// </summary>
-		Error
-	}
-
-	/// <summary>
-	/// Defines the presentation states for content sharing on a Cisco codec.
-	/// </summary>
-	public enum eCodecPresentationStates
-	{
-		/// <summary>
-		/// Presentation content is displayed locally only
-		/// </summary>
-		LocalOnly,
-		/// <summary>
-		/// Presentation content is displayed both locally and remotely
-		/// </summary>
-		LocalRemote
-	}
-
-	/// <summary>
-	/// Defines the camera tracking capabilities available on a Cisco codec.
-	/// </summary>
-	public enum eCameraTrackingCapabilities
-	{
-		/// <summary>
-		/// No camera tracking capabilities available
-		/// </summary>
-		None,
-		/// <summary>
-		/// SpeakerTrack capability available - automatically tracks active speakers
-		/// </summary>
-		SpeakerTrack,
-		/// <summary>
-		/// PresenterTrack capability available - automatically tracks presenters
-		/// </summary>
-		PresenterTrack,
-		/// <summary>
-		/// Both SpeakerTrack and PresenterTrack capabilities available
-		/// </summary>
-		Both
-	}
-
-	#endregion
 
 	/// <summary>
 	/// Extended Cisco video codec implementation that provides comprehensive control and monitoring capabilities for Cisco Room OS devices.
@@ -157,7 +51,7 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 	/// - Occupancy and people count monitoring
 	/// - WebEx integration and branding
 	/// </remarks>
-	public class CiscoCodec
+	public partial class CiscoCodec
 		: VideoCodecBase,
 			IHasCallHistory,
 			IHasCallFavorites,
@@ -214,21 +108,6 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 		private bool _IsInPresentation;
 
 		private MediaChannelStatus _incomingPresentation;
-
-		[Flags]
-		public enum MediaChannelStatus
-		{
-			Unknown = 0,
-			None = 1,
-			Outgoing = 2,
-			Incoming = 4,
-			Video = 8,
-			Audio = 16,
-			Main = 32,
-			Presentation = 64,
-			OutgoingPresentation = 66,
-			IncomingPresentation = 68
-		}
 
 		private readonly Version _testedCodecFirmware = new Version("10.11.5.2");
 		private readonly Version _enhancedLayoutsFirmware = new Version("9.15.10.8");
@@ -650,7 +529,7 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 			if (CodecFirmware == null)
 				return false;
 			var returnValue = CodecFirmware.CompareTo(ver) >= 0;
-			return (returnValue);
+			return returnValue;
 		}
 
 		#region CameraAutoTrackingFeedbackFunc
@@ -696,7 +575,7 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 			get
 			{
 				return () =>
-					((PresenterTrackStatus) || (String.IsNullOrEmpty(PresenterTrackStatusName)));
+					PresenterTrackStatus || string.IsNullOrEmpty(PresenterTrackStatusName);
 			}
 		}
 
@@ -785,6 +664,8 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 
 		private readonly IBasicCommunication _comms;
 
+		private EMonitorRole selfViewDefaultMonitorRole; // Default monitor role for self view, set in config
+
 		// Constructor for IBasicCommunication
 		public CiscoCodec(DeviceConfig config, IBasicCommunication comm)
 			: base(config)
@@ -801,15 +682,12 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 			UIExtensionsHandler = new UIExtensionsHandler(this, comm, _receiveQueue);
 			_comms = comm;
 
-
-
-
 			CrestronEnvironment.ProgramStatusEventHandler += a =>
-{
-	if (a != eProgramStatusEventType.Stopping)
-		return;
-	EndGracefully();
-};
+			{
+				if (a != eProgramStatusEventType.Stopping)
+					return;
+				EndGracefully();
+			};
 
 			CrestronEnvironment.SystemEventHandler += a =>
 			{
@@ -4795,7 +4673,7 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 			{
 				return obj["ResultId"].ToString();
 			}
-			catch (Exception ex)
+			catch
 			{
 				return Guid.Empty.ToString();
 			}
@@ -4825,640 +4703,6 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 						);
 					}
 				}
-
-				#region status
-				/*
-                if (response.IndexOf("\"Status\":{") > -1 || response.IndexOf("\"Status\": {") > -1)
-                {
-                    return;
-                    // Status DiagnosticsMessage
-
-                    // Temp object so we can inpsect for call data before simply deserializing
-                    var tempCodecStatus = new CiscoCodecStatus.RootObject();
-
-                    JsonConvert.PopulateObject(response, tempCodecStatus);
-
-                    var status = tempCodecStatus.Status;
-
-
-                    if (status.SystemUnit != null)
-                    {
-                        JsonConvert.PopulateObject(response, CodecStatus.Status.SystemUnit);
-                    }
-                    if (status.Cameras != null)
-                    {
-                        JsonConvert.PopulateObject(response, CodecStatus.Status.Cameras);
-                    }
-
-
-                    var network = status.Networks;
-
-                    if (network != null)
-                    {
-                        var myNetwork = network.FirstOrDefault(i => i.NetworkId == "1");
-                        if (myNetwork != null)
-                        {
-                            var hostname = myNetwork.Cdp.DeviceId.Value.NullIfEmpty() ?? "Unknown";
-                            var ipAddress = myNetwork.IPv4.Address.Value.NullIfEmpty() ?? "Unknown";
-                            var macAddress = myNetwork.Ethernet.MacAddress.Value.NullIfEmpty()
-                                             ?? "Unknown";
-
-
-                            OnCodecInfoChanged(new CodecInfoChangedEventArgs(eCodecInfoChangeType.Network)
-                            {
-                                IpAddress = ipAddress
-                            });
-
-                            if (DeviceInfo != null)
-                            {
-                                DeviceInfo.HostName = hostname;
-                                DeviceInfo.IpAddress = ipAddress;
-                                DeviceInfo.MacAddress = macAddress;
-                                UpdateDeviceInfo();
-                            }
-                        }
-                    }
-
-
-                    var sip = status.Sip;
-                    if (sip != null)
-                    {
-                        JsonConvert.PopulateObject(response, CodecStatus.Status.Sip);
-                        if (sip.Registrations.Count > 0)
-                        {
-                            var sipUri = sip.Registrations[0].Uri.Value.NullIfEmpty() ?? "Unknown";
-                            var match = Regex.Match(sipUri, @"(\d+)");
-                            var sipPhoneNumber = match.Success ? match.Groups[1].Value : "Unknown";
-                            OnCodecInfoChanged(new CodecInfoChangedEventArgs(eCodecInfoChangeType.Sip)
-                            {
-                                SipPhoneNumber = sipPhoneNumber,
-                                SipUri = sipUri
-                            });
-                        }
-                    }
-
-                    if (status.Video != null && status.Video.Layout != null)
-                    {
-                        if (status.Video.Layout.CurrentLayouts != null)
-                        {
-                            //TODO [ ] PREDICTION FIX
-                            //Debug.Console(0, this, "DeserializeVideo");
-                            JsonConvert.PopulateObject(response, CodecStatus.Status.Video.Layout.CurrentLayouts);
-                            if (status.Video.Layout.CurrentLayouts.AvailableLayouts != null)
-                            {
-                                UpdateLayoutList();
-                            }
-                            if (status.Video.Layout.CurrentLayouts.ActiveLayout == null) return;
-
-                            UpdateCurrentLayout();
-
-                        }
-                        else
-                        {
-                            CodecPollLayouts();
-                        }
-                    }
-
-
-                    // Check to see if the message contains /Status/ExperimentalConference/Presentation/PresentationLocalInstances and extract source value
-
-                    var conference = status.StatusConference;
-
-                    if (conference != null)
-                    {
-                        JsonConvert.PopulateObject(response, CodecStatus.Status.StatusConference);
-                    }
-
-                    if (conference != null &&
-                        (conference.Presentation != null && conference.Presentation.PresentationLocalInstances == null))
-                    {
-
-                        // Handles an empty presentation object response
-                        //return;
-                    }
-
-                    if (conference != null && conference.Presentation != null)
-                    {
-
-                        if (conference.Presentation.PresentationLocalInstances != null &&
-                            conference.Presentation.PresentationLocalInstances.Count > 0)
-                        {
-                            if (conference.Presentation.ModeValueProperty != null)
-                            {
-                                _presentationActive = conference.Presentation.ModeValueProperty.Value != "Off";
-                                CodecPollLayouts();
-                            }
-                            if (!string.IsNullOrEmpty(conference.Presentation.PresentationLocalInstances[0].Ghost))
-                            {
-                                _presentationSource = 0;
-                                _presentationLocalOnly = false;
-                                _presentationLocalRemote = false;
-                            }
-                            else if (conference.Presentation.PresentationLocalInstances[0].SourceValueProperty != null)
-                            {
-                                _presentationSource =
-                                    conference.Presentation.PresentationLocalInstances[0].SourceValueProperty.IntValue;
-
-                                // Check for any values in the SendingMode property
-                                if (
-                                    conference.Presentation.PresentationLocalInstances.Any(
-                                        (i) => !string.IsNullOrEmpty(i.SendingMode.Value)))
-                                {
-                                    _presentationLocalOnly =
-                                        conference.Presentation.PresentationLocalInstances.Any(
-                                            (i) => i.SendingMode.LocalOnly);
-                                    _presentationLocalRemote =
-                                        conference.Presentation.PresentationLocalInstances.Any(
-                                            (i) => i.SendingMode.LocalRemote);
-                                }
-                            }
-
-                            PresentationSourceFeedback.FireUpdate();
-                            PresentationSendingLocalOnlyFeedback.FireUpdate();
-                            PresentationSendingLocalRemoteFeedback.FireUpdate();
-                        }
-                    }
-
-                    var calls = status.Calls ?? null;
-
-                    // Check to see if this is a call status message received after the initial status message
-                    if (calls != null && calls.Count > 0)
-                    {
-                        if (calls.Count == 1 && !_presentationActive)
-                        {
-                            Debug.Console(0, this, "CALL WITH NO PRESENTATION!!!");
-                            ClearLayouts();
-                        }
-
-
-                        // Iterate through the call objects in the response
-                        foreach (var c in calls)
-                        {
-                            var call = c;
-
-                            var currentCallType = String.Empty;
-
-                            if (tempCodecStatus != null && tempCodecStatus.Status != null &&
-                                tempCodecStatus.Status.MediaChannels != null)
-                            {
-                                currentCallType = tempCodecStatus.Status.MediaChannels.MediaChannelCalls == null
-                                    ? null
-                                    : CheckCallType(c.CallIdString,
-                                        tempCodecStatus.Status.MediaChannels.MediaChannelCalls);
-                            }
-
-
-
-                            Debug.Console(0, this, "Current Call Type = {0}", currentCallType);
-                            var tempActiveCall = ActiveCalls.FirstOrDefault(x => x.Id.Equals(call.CallIdString));
-
-                            if (tempActiveCall != null)
-                            {
-
-                                var changeDetected = false;
-
-                                //var newStatus = eCodecCallStatus.Unknown;
-
-                                // Update properties of ActiveCallItem
-                                if (call.CallStatus != null)
-                                    if (!string.IsNullOrEmpty(call.CallStatus.Value))
-                                    {
-                                        tempActiveCall.Status =
-                                            CodecCallStatus.ConvertToStatusEnum(call.CallStatus.Value);
-                                        tempActiveCall.IsOnHold = tempActiveCall.Status == eCodecCallStatus.OnHold;
-
-                                        if (tempActiveCall.Status == eCodecCallStatus.Connected)
-                                        {
-
-                                            GetCallHistory();
-                                        }
-
-                                        changeDetected = true;
-                                    }
-
-                                if (call.CallType != null || currentCallType != null)
-                                {
-
-                                    tempActiveCall.Type =
-                                        CodecCallType.ConvertToTypeEnum(currentCallType ?? call.CallType.Value);
-                                    changeDetected = true;
-                                }
-
-
-
-                                if (call.DisplayName != null)
-                                    if (!string.IsNullOrEmpty(call.DisplayName.Value))
-                                    {
-                                        tempActiveCall.Name = call.DisplayName.Value;
-                                        changeDetected = true;
-                                    }
-
-
-                                if (call.Direction != null)
-                                {
-                                    if (!string.IsNullOrEmpty(call.Direction.Value))
-                                    {
-                                        tempActiveCall.Direction =
-                                            CodecCallDirection.ConvertToDirectionEnum(call.Direction.Value);
-                                        changeDetected = true;
-                                    }
-                                }
-                                if (call.Duration != null)
-                                {
-                                    if (!string.IsNullOrEmpty(call.Duration.Value))
-                                    {
-                                        tempActiveCall.Duration = call.Duration.DurationValue;
-                                        changeDetected = true;
-                                    }
-                                }
-                                if (call.PlacedOnHold != null)
-                                {
-
-
-                                    tempActiveCall.IsOnHold = call.PlacedOnHold.BoolValue;
-                                    changeDetected = true;
-                                }
-
-                                if (!changeDetected) continue;
-
-                                SetSelfViewMode();
-                                OnCallStatusChange(tempActiveCall);
-                                ListCalls();
-                                CodecPollLayouts();
-
-                            }
-                            else if (call.GhostString == null || call.GhostString.ToLower() == "false")
-                                // if the ghost value is present the call has ended already
-                            {
-
-                                // Create a new call item
-                                var newCallItem = new CodecActiveCallItem()
-                                {
-
-                                    Id = call.CallIdString,
-                                    Status = CodecCallStatus.ConvertToStatusEnum(call.CallStatus.Value),
-                                    Name = call.DisplayName.Value,
-                                    Number = call.RemoteNumber.Value,
-                                    Type = CodecCallType.ConvertToTypeEnum(currentCallType ?? call.CallType.Value),
-                                    Direction = CodecCallDirection.ConvertToDirectionEnum(call.Direction.Value),
-                                    Duration = call.Duration.DurationValue,
-                                    IsOnHold = call.PlacedOnHold.BoolValue,
-                                };
-
-
-                                // Add it to the ActiveCalls List
-                                ActiveCalls.Add(newCallItem);
-
-                                ListCalls();
-
-                                SetSelfViewMode();
-
-                                OnCallStatusChange(newCallItem);
-
-                                CodecPollLayouts();
-
-                                //ClearLayouts();
-
-                            }
-
-                        }
-
-                    }
-                    if (status.Audio != null)
-                    {
-                        JsonConvert.PopulateObject(response, CodecStatus.Status.Audio);
-                    }
-
-
-                    // Check for Room Preset data (comes in partial, so we need to handle these responses differently to prevent appending duplicate items
-                    var tempPresets = tempCodecStatus.Status.RoomPresets;
-
-                    if (tempPresets.Count > 0)
-                    {
-                        // Create temporary list to store the existing items from the CiscoCodecStatus.RoomPreset collection
-                        var existingRoomPresets = new List<CiscoCodecStatus.RoomPreset>();
-                        // Add the existing items to the temporary list
-                        existingRoomPresets.AddRange(CodecStatus.Status.RoomPresets);
-                        // Populate the CodecStatus object (this will append new values to the RoomPreset collection
-                        JsonConvert.PopulateObject(response, CodecStatus);
-
-                        var jResponse = JObject.Parse(response);
-
-
-                        IList<JToken> roomPresets = jResponse["Status"]["RoomPreset"].Children().ToList();
-                        // Iterate the new items in this response agains the temporary list.  Overwrite any existing items and add new ones.
-                        foreach (var camPreset in tempPresets)
-                        {
-                            var preset = camPreset;
-                            if (preset == null) continue;
-                            // First fine the existing preset that matches the CiscoCallId
-                            var existingPreset =
-                                existingRoomPresets.FirstOrDefault(p => p.RoomPresetId.Equals(preset.RoomPresetId));
-                            if (existingPreset != null)
-                            {
-                                Debug.Console(1, this, "Existing Room Preset with ID: {0} found. Updating.",
-                                    existingPreset.RoomPresetId);
-
-                                JToken updatedPreset = null;
-
-                                // Find the JToken from the response with the matching CiscoCallId
-                                foreach (
-                                    var jPreset in
-                                        roomPresets.Where(
-                                            jPreset =>
-                                                jPreset["id"].Value<string>() == existingPreset.RoomPresetId)
-                                    )
-                                {
-                                    updatedPreset = jPreset;
-                                }
-
-                                if (updatedPreset != null)
-                                {
-                                    // use PopulateObject to overlay the partial data onto the existing object
-                                    JsonConvert.PopulateObject(updatedPreset.ToString(), existingPreset);
-                                }
-
-                            }
-                            else
-                            {
-                                Debug.Console(1, this, "New Room Preset with ID: {0}. Adding.", preset.RoomPresetId);
-                                existingRoomPresets.Add(preset);
-                            }
-                        }
-
-                        // Replace the list in the CodecStatus object with the processed list
-                        CodecStatus.Status.RoomPresets = existingRoomPresets;
-
-                        // Generecise the list
-                        NearEndPresets =
-                            existingRoomPresets.GetGenericPresets<CiscoCodecStatus.RoomPreset, CodecRoomPreset>();
-
-                        var handler = CodecRoomPresetsListHasChanged;
-                        if (handler != null)
-                        {
-                            handler(this, new EventArgs());
-                        }
-                    }
-
-                    else
-                    {
-                        JsonConvert.PopulateObject(response, CodecStatus);
-                    }
-
-
-                    if (_syncState.InitialStatusMessageWasReceived) return;
-                    _syncState.InitialStatusMessageReceived();
-
-                    if (!_syncState.InitialConfigurationMessageWasReceived)
-                    {
-                        Debug.Console(0, this, "Sending Configuration");
-                        SendText("xConfiguration");
-                    }
-                    if (!_syncState.FeedbackWasRegistered)
-                    {
-                        Debug.Console(0, this, "Sending Feedback");
-
-                        SendText(BuildFeedbackRegistrationExpression());
-                    }
-                }
-
-                #endregion
-
-                #region Configuration
-
-                /*
-        if (response.IndexOf("\"Configuration\":{") > -1 || response.IndexOf("\"Configuration\": {") > -1)
-
-                {
-                    // Configuration DiagnosticsMessage
-                    Debug.Console(2, this, "Parse Configuration : {0}", response);
-
-                    JsonConvert.PopulateObject(response, CodecConfiguration);
-
-                    if (CodecConfiguration != null)
-                    {
-                        var h323 = CodecConfiguration.Configuration.H323;
-                        if (h323 != null)
-                        {
-                            if (h323.H323Alias != null)
-                            {
-                                var e164 = h323.H323Alias.E164.Value.NullIfEmpty() ?? "unknown";
-                                var h323Id = h323.H323Alias.H323AliasId.Value.NullIfEmpty() ?? "unknown";
-                                OnCodecInfoChanged(new CodecInfoChangedEventArgs(eCodecInfoChangeType.H323)
-                                {
-                                    E164Alias = e164,
-                                    H323Id = h323Id
-                                });
-                            }
-                        }
-                        bool autoAnswer;
-                        if (CodecConfiguration.Configuration.Conference.AutoAnswer.AutoAnswerMode.Value == null)
-                            autoAnswer = false;
-                        else
-                        {
-                            autoAnswer =
-                                CodecConfiguration.Configuration.Conference.AutoAnswer.AutoAnswerMode.Value.ToLower() == "on";
-                        }
-                        OnCodecInfoChanged(new CodecInfoChangedEventArgs(eCodecInfoChangeType.AutoAnswer)
-                        {
-                            AutoAnswerEnabled = autoAnswer
-                        });
-                    }
-
-                    if (!_syncState.InitialConfigurationMessageWasReceived)
-                    {
-                        Debug.Console(2, this, "InitialConfig Received");
-                        _syncState.InitialConfigurationMessageReceived();
-                        if (!_syncState.InitialSoftwareVersionMessageWasReceived)
-                        {
-                            SendText("xStatus SystemUnit");
-                        }
-                    }
-
-                }
-         */
-
-				#endregion
-
-				#region event
-
-				/*
-        if (response.IndexOf("\"Event\":{") > -1 || response.IndexOf("\"Event\": {") > -1)
-                {
-                    Debug.Console(2, this, "Parse Event : {0}", response);
-                    if (response.IndexOf("\"CallDisconnect\":{") > -1 || response.IndexOf("\"CallDisconnect\": {") > -1)
-                    {
-                        Debug.Console(2, this, "Parse CallDisconnect");
-                        var eventReceived = new CiscoCodecEvents.Event();
-
-                        JsonConvert.PopulateObject(response, eventReceived);
-
-                        EvalutateDisconnectEvent(eventReceived);
-                    }
-                    else if (response.IndexOf("\"Bookings\":{") > -1 || response.IndexOf("\"Bookings\": {") > -1)
-                        // The list has changed, reload it
-                    {
-                        Debug.Console(2, this, "Parse Bookings");
-
-                        GetBookings(null);
-                    }
-
-                    else if (response.IndexOf("\"UserInterface\":{") > -1 ||
-                             response.IndexOf("\"UserInterface\": {") > -1) // External SourceValueProperty Trigger
-                    {
-                        Debug.Console(2, this, "Parse UserInterface");
-
-                        var eventReceived = new CiscoCodecEvents.RootObject();
-                        JsonConvert.PopulateObject(response, eventReceived);
-                        Debug.Console(2, this, "*** Got an External SourceValueProperty Selection {0} {1}",
-                            eventReceived,
-                            eventReceived.Event.UserInterface,
-                            eventReceived.Event.UserInterface.Presentation.ExternalSource.Selected
-                                .SourceIdentifier.Value);
-
-                        if (RunRouteAction != null && !_externalSourceChangeRequested)
-                        {
-                            RunRouteAction(
-                                eventReceived.Event.UserInterface.Presentation.ExternalSource.Selected
-                                    .SourceIdentifier.Value, null);
-                        }
-
-                        _externalSourceChangeRequested = false;
-                    }
-                }
-         */
-
-				#endregion
-
-				#region commandResponse
-
-				/*
-        else if (response.IndexOf("\"CommandResponse\":{") > -1 ||
-                         response.IndexOf("\"CommandResponse\": {") > -1)
-                {
-                    // CommandResponse DiagnosticsMessage
-                    Debug.Console(2, this, "Parse CommandResponse - {0}", response);
-
-
-                    if (response.IndexOf("\"CallHistoryRecentsResult\":{") > -1 ||
-                        response.IndexOf("\"CallHistoryRecentsResult\": {") > -1)
-                    {
-                        Debug.Console(2, this, "Parse CallHistoryRecentsResult");
-
-                        var codecCallHistory = new CiscoCallHistory.RootObject();
-
-                        JsonConvert.PopulateObject(response, codecCallHistory);
-
-                        CallHistory.ConvertCiscoCallHistoryToGeneric(
-                            codecCallHistory.CommandResponse.CallHistoryRecentsResult.Entry);
-                    }
-                    else if (response.IndexOf("\"CallHistoryDeleteEntryResult\":{") > -1 ||
-                             response.IndexOf("\"CallHistoryDeleteEntryResult\": {") > -1)
-                    {
-                        Debug.Console(2, this, "Parse GetCallHistoryDeleteEntryResult");
-
-                        GetCallHistory();
-                    }
-                    else if (response.IndexOf("\"PhonebookSearchResult\":{") > -1 ||
-                             response.IndexOf("\"PhonebookSearchResult\": {") > -1)
-                    {
-                        Debug.Console(2, this, "Parse PhonebookSearchResult");
-
-                        var codecPhonebookResponse = new CiscoCodecExtendedPhonebook.RootObject();
-
-                        JsonConvert.PopulateObject(response, codecPhonebookResponse);
-
-                        if (!PhonebookSyncState.InitialPhonebookFoldersWasReceived)
-                        {
-                            // Check if the phonebook has any folders
-                            PhonebookSyncState.InitialPhonebookFoldersReceived();
-
-                            PhonebookSyncState.SetPhonebookHasFolders(
-                                codecPhonebookResponse.CommandResponse.PhonebookSearchResult.Folder.Count >
-                                0);
-
-                            if (PhonebookSyncState.PhonebookHasFolders)
-                            {
-                                DirectoryRoot.AddFoldersToDirectory(
-                                    CiscoCodecExtendedPhonebook.GetRootFoldersFromSearchResult(
-                                        codecPhonebookResponse.CommandResponse.PhonebookSearchResult));
-                            }
-
-                            // Get the number of contacts in the phonebook
-                            GetPhonebookContacts();
-                        }
-                        else if (!PhonebookSyncState.NumberOfContactsWasReceived)
-                        {
-                            // Store the total number of contacts in the phonebook
-                            PhonebookSyncState.SetNumberOfContacts(
-                                Int32.Parse(
-                                    codecPhonebookResponse.CommandResponse.PhonebookSearchResult.ResultInfo
-                                        .TotalRows.Value));
-
-                            DirectoryRoot.AddContactsToDirectory(
-                                CiscoCodecExtendedPhonebook.GetRootContactsFromSearchResult(
-                                    codecPhonebookResponse.CommandResponse.PhonebookSearchResult));
-
-                            PhonebookSyncState.PhonebookRootEntriesReceived();
-
-                            PrintDirectory(DirectoryRoot);
-                        }
-                        else if (PhonebookSyncState.InitialSyncComplete)
-                        {
-                            var directoryResults = new CodecDirectory();
-
-                            if (
-                                codecPhonebookResponse.CommandResponse.PhonebookSearchResult.ResultInfo
-                                    .TotalRows.Value != "0")
-                                directoryResults =
-                                    CiscoCodecExtendedPhonebook.ConvertCiscoPhonebookToGeneric(
-                                        codecPhonebookResponse.CommandResponse.PhonebookSearchResult);
-
-                            PrintDirectory(directoryResults);
-
-                            DirectoryBrowseHistory.Add(directoryResults);
-
-                            OnDirectoryResultReturned(directoryResults);
-
-                        }
-                    }
-                    else if (response.IndexOf("\"BookingsListResult\":{") > -1
-                             || response.IndexOf("\"BookingsListResult\": {") > -1)
-                    {
-                        Debug.Console(2, this, "Parse BookingsListResult - {0}", response);
-
-                        var codecBookings = new CiscoExtendedCodecBookings.RootObject();
-
-                        JsonConvert.PopulateObject(response, codecBookings);
-
-                        if (codecBookings.CommandResponse.BookingsListResult.ResultInfo.TotalRows.Value !=
-                            "0")
-                        {
-                            Debug.Console(2, this, "THere are {0} meetings",
-                                codecBookings.CommandResponse.BookingsListResult.ResultInfo.TotalRows.Value);
-                            CodecSchedule.Meetings =
-                                CiscoExtendedCodecBookings.GetGenericMeetingsFromBookingResult(
-                                    codecBookings.CommandResponse.BookingsListResult.BookingsListResultBooking,
-                                    _joinableCooldownSeconds);
-                        }
-                        else
-                        {
-                            Debug.Console(2, this, "There are No Meetings");
-                            CodecSchedule.Meetings = new List<Meeting>();
-                        }
-                        if (BookingsRefreshTimer == null)
-                        {
-                            BookingsRefreshTimer = new CTimer(GetBookings, 90000, 90000);
-                            Debug.Console(2, this, "BookingsRefresh Was null");
-                        }
-                        BookingsRefreshTimer.Reset(90000, 90000);
-
-                    }
-
-                }
-             */
-
-				#endregion
 			}
 			catch (Exception ex)
 			{
@@ -5467,10 +4711,6 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 				if (ex is JsonReaderException)
 				{
 					Debug.Console(1, this, "Received malformed response from codec.");
-
-					//Communication.Disconnect();
-
-					//Initialize();
 				}
 			}
 		}
@@ -7740,240 +6980,4 @@ namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
 			EnqueueCommand($"xCommand UserInterface WebView Clear Target:OSD{CiscoCodec.Delimiter}");
 		}
 	}
-
-	#region
-
-	public class CodecInfoChangedEventArgs : EventArgs
-	{
-		public bool MultiSiteOptionIsEnabled { get; set; }
-		public string IpAddress { get; set; }
-		public string SipPhoneNumber { get; set; }
-		public string E164Alias { get; set; }
-		public string H323Id { get; set; }
-		public string SipUri { get; set; }
-		public bool AutoAnswerEnabled { get; set; }
-		public string Firmware { get; set; }
-		public string SerialNumber { get; set; }
-
-		public eCodecInfoChangeType InfoChangeType { get; private set; }
-
-		public CodecInfoChangedEventArgs()
-		{
-			InfoChangeType = eCodecInfoChangeType.Unknown;
-		}
-
-		public CodecInfoChangedEventArgs(eCodecInfoChangeType changeType)
-		{
-			InfoChangeType = changeType;
-		}
-	}
-
-	public enum eCodecInfoChangeType
-	{
-		AutoAnswer,
-		Network,
-		Sip,
-		H323,
-		Multisite,
-		Unknown,
-		SerialNumber,
-		Firmware
-	}
-
-	public class FeedbackGroup
-	{
-		private readonly FeedbackCollection<Feedback> _feedbacks;
-
-		public FeedbackGroup(IEnumerable<Feedback> feedbacks)
-		{
-			_feedbacks = new FeedbackCollection<Feedback>();
-			_feedbacks.AddRange(feedbacks);
-		}
-
-		public void FireUpdate()
-		{
-			foreach (var f in _feedbacks)
-			{
-				var feedback = f;
-				feedback.FireUpdate();
-			}
-		}
-	}
-
-	public class MinuteChangedEventArgs : EventArgs
-	{
-		public DateTime EventTime { get; private set; }
-
-		public MinuteChangedEventArgs(DateTime eventTime)
-		{
-			EventTime = eventTime;
-		}
-
-		public MinuteChangedEventArgs()
-		{
-			EventTime = DateTime.Now;
-		}
-	}
-
-	public class CameraTrackingCapabilitiesArgs : EventArgs
-	{
-		public eCameraTrackingCapabilities CameraTrackingCapabilites { get; set; }
-
-		public CameraTrackingCapabilitiesArgs(bool speakerTrack, bool presenterTrack)
-		{
-			CameraTrackingCapabilites = SetCameraTrackingCapabilities(speakerTrack, presenterTrack);
-		}
-
-		public CameraTrackingCapabilitiesArgs(Func<bool> speakerTrack, Func<bool> presenterTrack)
-		{
-			CameraTrackingCapabilites = SetCameraTrackingCapabilities(
-				speakerTrack(),
-				presenterTrack()
-			);
-		}
-
-		private eCameraTrackingCapabilities SetCameraTrackingCapabilities(
-			bool speakerTrack,
-			bool presenterTrack
-		)
-		{
-			var trackingType = eCameraTrackingCapabilities.None;
-
-			if (speakerTrack && presenterTrack)
-			{
-				trackingType = eCameraTrackingCapabilities.Both;
-				return trackingType;
-			}
-			if (!speakerTrack && presenterTrack)
-			{
-				trackingType = eCameraTrackingCapabilities.PresenterTrack;
-				return trackingType;
-			}
-			if (speakerTrack && !presenterTrack)
-			{
-				trackingType = eCameraTrackingCapabilities.SpeakerTrack;
-				return trackingType;
-			}
-			return trackingType;
-		}
-	}
-
-	/// <summary>
-	/// Provides device information and capabilities for a Cisco codec, extending the base VideoCodecInfo class.
-	/// This class contains codec-specific information such as multi-site capabilities, IP address, and other device details.
-	/// </summary>
-	public class CiscoCodecInfo : VideoCodecInfo
-	{
-		private readonly CiscoCodec _codec;
-
-		private bool _multiSiteOptionIsEnabled;
-
-		/// <inheritdoc />
-		public override bool MultiSiteOptionIsEnabled
-		{
-			get { return _multiSiteOptionIsEnabled; }
-		}
-
-		private string _ipAddress;
-
-		/// <inheritdoc />
-		public override string IpAddress
-		{
-			get { return _ipAddress; }
-		}
-
-		private string _sipPhoneNumber;
-
-		/// <inheritdoc />
-		public override string SipPhoneNumber
-		{
-			get { return _sipPhoneNumber; }
-		}
-
-		private string _e164Alias;
-
-		/// <inheritdoc />
-		public override string E164Alias
-		{
-			get { return _e164Alias; }
-		}
-
-		private string _h323Id;
-
-		public override string H323Id
-		{
-			get { return _h323Id; }
-		}
-
-		private string _sipUri;
-
-		public override string SipUri
-		{
-			get { return _sipUri; }
-		}
-
-		private bool _autoAnswerEnabled;
-
-		public override bool AutoAnswerEnabled
-		{
-			get { return _autoAnswerEnabled; }
-		}
-
-		public CiscoCodecInfo(CiscoCodec codec)
-		{
-			_codec = codec;
-			_codec.CodecInfoChanged += (sender, args) =>
-			{
-				if (args.InfoChangeType == eCodecInfoChangeType.Unknown)
-					return;
-				switch (args.InfoChangeType)
-				{
-					case eCodecInfoChangeType.Network:
-						_ipAddress = args.IpAddress;
-						break;
-					case eCodecInfoChangeType.Sip:
-						_sipPhoneNumber = args.SipPhoneNumber;
-						_sipUri = args.SipUri;
-						break;
-					case eCodecInfoChangeType.H323:
-						_h323Id = args.H323Id;
-						_e164Alias = args.E164Alias;
-						break;
-					case eCodecInfoChangeType.Multisite:
-						_multiSiteOptionIsEnabled = args.MultiSiteOptionIsEnabled;
-						break;
-					case eCodecInfoChangeType.AutoAnswer:
-						_autoAnswerEnabled = args.AutoAnswerEnabled;
-						break;
-				}
-			};
-		}
-	}
-
-	public static class ExtensionsMethods
-	{
-		public static string EncodeBase64(this string plainText)
-		{
-			var plainTextBytes = Encoding.UTF8.GetBytes(plainText);
-			return Convert.ToBase64String(plainTextBytes);
-		}
-
-		public static string DecodeBase64(this string encodedText)
-		{
-			var encodedTextBytes = Encoding.UTF8.GetBytes(encodedText);
-			return Convert.ToString(encodedTextBytes);
-		}
-
-		public static JToken GetValueProperty(this JObject jObject, string path)
-		{
-			JToken outToken;
-
-			if (!jObject.TryGetValue(path, out outToken))
-				return null;
-			var outputValue = outToken.SelectToken("Value");
-			return outputValue;
-		}
-	}
-
-	#endregion
 }
