+++ src/EMonitorRole.cs
@@ -1,0 +1,10 @@
+namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
+{
+  public enum EMonitorRole
+  {
+    First,
+    Second,
+
+    Third,
+  }
+}
