+++ src/Enums/eCodecPresentationStates.cs
@@ -1,0 +1,18 @@
+namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec.Enums
+{
+  /// <summary>
+  /// Defines the presentation states for content sharing on a Cisco codec.
+  /// </summary>
+  public enum eCodecPresentationStates
+  {
+    /// <summary>
+    /// Presentation content is displayed locally only
+    /// </summary>
+    LocalOnly,
+    /// <summary>
+    /// Presentation content is displayed both locally and remotely
+    /// </summary>
+    LocalRemote
+  }
+
+}
