+++ src/Enums/eExternalSourceType.cs
@@ -1,0 +1,37 @@
+namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec.Enums
+{
+  /// <summary>
+  /// Defines the types of external sources that can be connected to a Cisco codec.
+  /// </summary>
+  public enum eExternalSourceType
+  {
+    /// <summary>
+    /// External camera source
+    /// </summary>
+    camera,
+    /// <summary>
+    /// Desktop computer source
+    /// </summary>
+    desktop,
+    /// <summary>
+    /// Document camera source
+    /// </summary>
+    document_camera,
+    /// <summary>
+    /// Media player source
+    /// </summary>
+    mediaplayer,
+    /// <summary>
+    /// Personal computer source
+    /// </summary>
+    PC,
+    /// <summary>
+    /// Whiteboard source
+    /// </summary>
+    whiteboard,
+    /// <summary>
+    /// Other unspecified source type
+    /// </summary>
+    other
+  }
+}
