+++ src/eCodecInfoChangeType.cs
@@ -1,0 +1,18 @@
+namespace PepperDash.Essentials.Plugin.CiscoRoomOsCodec
+{
+  #region
+
+  public enum eCodecInfoChangeType
+  {
+    AutoAnswer,
+    Network,
+    Sip,
+    H323,
+    Multisite,
+    Unknown,
+    SerialNumber,
+    Firmware
+  }
+
+  #endregion
+}
